organize into cdc and qrep folders
